refactor: 프로필 nav 하드 텍스트 제거하고 QUERY_PARAMS 상수를 url.ts 파일로 이동 (#124)
