--- src/constant/queryParams.ts
@@ -1,7 +1,0 @@
-export const QUERY_PARAMS = {
-  PLAYLIST_SN: 'playlistSn',
-  PLAYLIST_INFO: 'playlistInfo',
-  PLAYLIST_INDEX: 'playlistIndex',
-  PLAYLIST_COMMENTS: 'comments',
-  VIDEO_INDEX: 'videoIndex',
-};
